requirements.txt
openai>=1.51.0
pillow>=10.0.0
python-dotenv>=1.0.1
streamlit>=1.33.0
